Redesign API of ExpectationMaximization
Previously the EM-related classes defined the data they would
operate on during constructions, and functions performing the
calculations where parameterless and void.

Now the constructors are more vague and treat parameters as a
way to initialize their internal cache-helpers, while functions
accept all the data they work on as parameters.